+++ src/main/java/site/offload/offloadserver/db/announcement/repository/AnnouncementRepository.java
@@ -1,0 +1,7 @@
+package site.offload.offloadserver.db.announcement.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import site.offload.offloadserver.db.announcement.entity.Announcement;
+
+public interface AnnouncementRepository extends CrudRepository<Announcement, Long> {
+}
