+++ src/main/java/site/offload/offloadserver/db/character/repository/CharacterRepository.java
@@ -1,0 +1,7 @@
+package site.offload.offloadserver.db.character.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import site.offload.offloadserver.db.character.entity.Character;
+
+public interface CharacterRepository extends CrudRepository<Character, Integer> {
+}
