+++ src/main/java/site/offload/offloadserver/db/character/repository/GainedCharacterRepository.java
@@ -1,0 +1,7 @@
+package site.offload.offloadserver.db.character.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import site.offload.offloadserver.db.character.entity.GainedCharacter;
+
+public interface GainedCharacterRepository extends CrudRepository<GainedCharacter, Long> {
+}
