+++ src/main/java/site/offload/offloadserver/db/charactermotion/entity/CharacterMotion.java
@@ -1,0 +1,29 @@
+package site.offload.offloadserver.db.charactermotion.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import site.offload.offloadserver.db.character.entity.Character;
+import site.offload.offloadserver.db.BaseTimeEntity;
+import site.offload.offloadserver.db.place.entity.PlaceCategory;
+
+//각 캐릭터의 모션
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class CharacterMotion extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "character_id", nullable = false)
+    private Character character;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private PlaceCategory placeCategory;
+
+    @Column(nullable = false, unique = true, columnDefinition = "TEXT")
+    private String motionImageUrl;
+}
