+++ src/main/java/site/offload/offloadserver/db/emblem/repository/GainedEmblemRepository.java
@@ -1,0 +1,7 @@
+package site.offload.offloadserver.db.emblem.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import site.offload.offloadserver.db.emblem.entity.GainedEmblem;
+
+public interface GainedEmblemRepository extends CrudRepository<GainedEmblem, Integer> {
+}
