--- src/main/java/site/offload/offloadserver/db/entity/Announcement.java
@@ -1,25 +1,0 @@
-package site.offload.offloadserver.db.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-//공지사항
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Announcement extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String writer;
-
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String title;
-
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String content;
-}
