--- src/main/java/site/offload/offloadserver/db/entity/Character.java
@@ -1,28 +1,0 @@
-package site.offload.offloadserver.db.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-//캐릭터
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Character extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String description;
-
-    @Column(nullable = false, unique = true, columnDefinition = "TEXT")
-    private String characterBaseImageUrl;
-
-    @Column(nullable = false, unique = true)
-    private String CharacterCode;
-}
