--- src/main/java/site/offload/offloadserver/db/entity/Emblem.java
@@ -1,11 +1,0 @@
-package site.offload.offloadserver.db.entity;
-
-import lombok.RequiredArgsConstructor;
-
-//칭호
-@RequiredArgsConstructor
-public enum Emblem  {
-    //예시 칭호
-    EMBLEM_CODE("EMBLEM001");
-    private final String emblemCodeName;
-}
