--- src/main/java/site/offload/offloadserver/db/entity/MemberGender.java
@@ -1,7 +1,0 @@
-package site.offload.offloadserver.db.entity;
-
-//유저 성별
-public enum MemberGender {
-    MALE,
-    FEMALE
-}
