--- src/main/java/site/offload/offloadserver/db/entity/Place.java
@@ -1,35 +1,0 @@
-package site.offload.offloadserver.db.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-//서비스에 등록된 장소
-@Entity
-@NoArgsConstructor(access =  AccessLevel.PROTECTED)
-public class Place extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String address;
-
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String shortIntroduction;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private PlaceCategory placeCategory;
-
-    @Column(nullable = false, columnDefinition = "float CHECK (latitude >= -90 AND latitude <= 90)")
-    private float latitude;
-
-    @Column(nullable = false, columnDefinition = "float CHECK (longitude >= -180 AND longitude <= 180)")
-    private float longitude;
-}
