--- src/main/java/site/offload/offloadserver/db/entity/PlaceCategory.java
@@ -1,11 +1,0 @@
-package site.offload.offloadserver.db.entity;
-
-//장소 카테고리
-public enum PlaceCategory {
-    CAFFE,
-    PARK,
-    RESTAURANT,
-    CULTURE,
-    SPORT,
-    NONE
-}
