--- src/main/java/site/offload/offloadserver/db/entity/Quest.java
@@ -1,8 +1,0 @@
-package site.offload.offloadserver.db.entity;
-
-public enum Quest {
-
-    //예시 퀘스트 코드
-    QUEST_CODE;
-    private String QuestCode;
-}
