+++ src/main/java/site/offload/offloadserver/db/member/entity/MemberGender.java
@@ -1,0 +1,7 @@
+package site.offload.offloadserver.db.member.entity;
+
+//유저 성별
+public enum MemberGender {
+    MALE,
+    FEMALE
+}
