+++ src/main/java/site/offload/offloadserver/db/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package site.offload.offloadserver.db.member.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import site.offload.offloadserver.db.member.entity.Member;
+
+public interface MemberRepository extends CrudRepository<Member, Long> {
+}
