+++ src/main/java/site/offload/offloadserver/db/place/entity/PlaceCategory.java
@@ -1,0 +1,11 @@
+package site.offload.offloadserver.db.place.entity;
+
+//장소 카테고리
+public enum PlaceCategory {
+    CAFFE,
+    PARK,
+    RESTAURANT,
+    CULTURE,
+    SPORT,
+    NONE
+}
