+++ src/main/java/site/offload/offloadserver/db/place/repository/VisitedPlaceRepositoiry.java
@@ -1,0 +1,7 @@
+package site.offload.offloadserver.db.place.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import site.offload.offloadserver.db.place.entity.VisitedPlace;
+
+public interface VisitedPlaceRepositoiry extends CrudRepository<VisitedPlace, Long> {
+}
