+++ src/main/java/site/offload/offloadserver/db/quest/entity/Quest.java
@@ -1,0 +1,8 @@
+package site.offload.offloadserver.db.quest.entity;
+
+public enum Quest {
+
+    //예시 퀘스트 코드
+    QUEST_CODE;
+    private String QuestCode;
+}
