--- src/main/java/site/offload/offloadserver/db/repository/GainedEmblemRepository.java
@@ -1,7 +1,0 @@
-package site.offload.offloadserver.db.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import site.offload.offloadserver.db.entity.GainedEmblem;
-
-public interface GainedEmblemRepository extends CrudRepository<GainedEmblem, Integer> {
-}
