--- src/main/java/site/offload/offloadserver/db/repository/PlaceRepository.java
@@ -1,7 +1,0 @@
-package site.offload.offloadserver.db.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import site.offload.offloadserver.db.entity.Place;
-
-public interface PlaceRepository extends CrudRepository<Place, Long> {
-}
